+++ CHANGELOG.md
@@ -20,6 +20,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 - `promote`, `embed` and `permute` for promoting, embeding and permuting quantum objects.
 
+### Changed
+
+- `plot` on `SuperOp` now plots in the computational basis.
+
 ## [0.6.3] - 2026-05-15
 
 ### Added